refactoring after upgrading discord.py
